Get rid of AbstractManagedExecutorServices - instead make AbstractExecutorServices more easily overridable
